feat: Add button to return to waiting room in Navbar
